Changing names of PICRUSt2 scripts that run using the previous database
